Renamed "impl" to "internal"
"internal" is what we use in the Node client and seems to better convey "Keep out".
